+++ packages/vue/src/components/field/examples/textarea.vue
@@ -5,7 +5,7 @@ import { Field } from '@ark-ui/vue/field'
 <template>
   <Field.Root>
     <Field.Label>Label</Field.Label>
-    <Field.Textarea />
+    <Field.Textarea autoresize />
     <Field.HelperText>Some additional Info</Field.HelperText>
     <Field.ErrorText>Error Info</Field.ErrorText>
   </Field.Root>