+++ packages/vue/src/components/select/examples/basic.vue
@@ -21,13 +21,7 @@ const collection = createListCollection({
     </Select.Control>
     <Teleport to="body">
       <Select.Positioner>
-        <Select.Content
-          :ref="
-            (el) => {
-              console.log(el)
-            }
-          "
-        >
+        <Select.Content>
           <Select.ItemGroup>
             <Select.ItemGroupLabel>Frameworks</Select.ItemGroupLabel>
             <Select.Item v-for="item in collection.items" :key="item" :item="item">