+++ packages/vue/src/components/select/examples/root-provider.vue
@@ -25,13 +25,7 @@ const select = useSelect({ collection: collection })
     </Select.Control>
     <Teleport to="body">
       <Select.Positioner>
-        <Select.Content
-          :ref="
-            (el) => {
-              console.log(el)
-            }
-          "
-        >
+        <Select.Content>
           <Select.ItemGroup>
             <Select.ItemGroupLabel>Frameworks</Select.ItemGroupLabel>
             <Select.Item v-for="item in collection.items" :key="item" :item="item">