+++ packages/vue/src/components/tree-view/tree-view-branch-content.vue
@@ -25,7 +25,6 @@ import { createSplitProps } from '../create-split-props'
 
 const splitVisibilityProps = createSplitProps<VisibilityProps>()
 
-
 defineProps<TreeViewBranchContentProps>()
 const treeView = useTreeViewContext()
 const ndoeProps = useTreeViewNodePropsContext()